Factor out password-reset directory.
